requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
autocommand==2.2.2
backcall==0.2.0
backoff==1.11.1
backports.tarfile==1.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.0.4
comm==0.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
filelock==3.7.1
frozenlist==1.5.0
fsspec==2025.3.0
gdown==3.13.1
googledrivedownloader==0.4
gql==3.4.0
graphql-core==3.1.4
h5py==3.4.0
hydra-core==1.0.7
idna==3.10
imageio==2.9.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
ipykernel==6.29.5
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.7.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.18.0
Jinja2==3.1.6
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab_widgets==3.0.15
kiwisolver==1.3.2
laspy==2.0.3
llvmlite==0.36.0
MarkupSafe==2.1.5
matplotlib==3.4.3
matplotlib-inline==0.1.3
MinkowskiEngine==0.5.4
mistune==0.8.4
more-itertools==10.3.0
multidict==6.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.10.4
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.11.1
numba==0.53.1
numpy==1.20.3
omegaconf==2.0.6
open3d==0.12.0
packaging==24.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==24.0
platformdirs==4.2.2
plyfile==0.7.4
promise==2.3
prompt-toolkit==3.0.20
propcache==0.2.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pybind11==2.13.6
pybind11_global==2.13.6
Pygments==2.10.0
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.8.2
python-louvain==0.16
pytorch-metric-learning==0.9.99
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-toolbelt==1.0.0
Rx==3.2.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==2.28.0
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
tensorboard==2.6.0
terminado==0.12.1
threadpoolctl==2.2.0
tomli==2.0.1
torch==1.12.1+cu113 # Core environment (strictly locked version)
torch_cluster==1.6.3
torch_geometric==1.7.2
torch-points-kernels==0.7.0
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==0.12.1+rocm5.1.1 # Core environment (strictly locked version)
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.13.1+cu113 # Core environment (strictly locked version)
tornado==6.1
tqdm==4.62.3
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.5.0
urllib3==2.2.3
visdom==0.1.8.9
wandb==0.8.36
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
wheel==0.44.0
yarl==1.15.2
zipp==3.19.2

